requirements.txt
# torch==1.13.0
# torch_complex==0.4.3
# torchaudio==0.13.0
# torchvision==0.14.0
# tqdm==4.65.0
# warpctc_pytorch==0.2.1
# av==9.1.0
# chainer==7.8.1
# ctcdecode==1.0.2
# editdistance==0.8.1
# gdown==4.6.0
# hydra-core==1.3.2
# ibug==0.0.9
# ibug_face_alignment.egg==info
# jiwer==1.2
# librosa==0.8.0
# matplotlib==3.6.2
# numpy==1.23.4
# omegaconf==2.3.0
# opencv_contrib_python==4.8.0.76
# Pillow==10.2.0
# requests==2.28.1
# scikit_image==0.19.2
# scipy==1.9.3
# sentencepiece==0.1.96
# six==1.16.0
# skimage==0.0
# soundfile==0.11.0

torch
torchaudio
torchvision
x-transformers
tensorboard
Unidecode
torchmetrics
pandas
inflect
g2p_en
transformers
openai-whisper
tqdm
chainer
# ctcdecode
editdistance
gdown
scoreq
# git+https://github.com/Takaaki-Saeki/DiscreteSpeechMetrics.git
# pip install https://github.com/vBaiCai/python-pesq/archive/master.zip
# git+https://github.com/WayenVan/ctcdecode.git#egg=ctcdecode install it after torch
hydra-core
jiwer
librosa
matplotlib
numpy<2.0.0
omegaconf
# opencv_contrib_python
# Pillow
requests
# scikit_image
scipy
sentencepiece
six
# skimage
soundfile
onnxruntime-gpu


